Split shell computer text input tests
